+++ cookbooks/base/recipes/main.rb
@@ -35,7 +35,7 @@ end
 package "dkms"
 
 # Adaptive readahead daemon
-package "preload"
+package "preload" if memory > 4.GB
 
 # NX client for QT
 package "qtnx"
@@ -47,4 +47,54 @@ package "p7zip-full"
 package "nautilus-filename-repairer"
 package "nautilus-gtkhash"
 package "nautilus-open-terminal"
+
+# Office productivity suite
+package "lo-menubar" if node[:platform_version] == "12.04"
+package "libreoffice-style-galaxy"
+
+# Commonly used restricted packages for Ubuntu
+package "ubuntu-restricted-extras" do
+  notifies :run, "execute[install_additional_extras]", :immediately
+end
+
+execute "install_additional_extras" do
+  cwd "/usr/share/doc/libdvdread4"
+  command "sh install-css.sh"
+  action :nothing
+end
+
+# MS Office True Type Fonts
+fonts = data_bag_item('resources', 'fonts')
+msfonts = fonts['msttfonts']
+
+fonts_file   = msfonts['file']
+fonts_url    = msfonts['url']
+fonts_sha256 = msfonts['sha256']
+
+cache_path = Chef::Config[:file_cache_path]
+fonts_path = "/usr/share/fonts/truetype/msttfonts"
+
+remote_file "#{cache_path}/#{fonts_file}" do
+  source fonts_url
+  checksum fonts_sha256
+  notifies :run, "bash[install_fonts]", :immediately
+end
+
+bash "install_fonts" do
+  cwd cache_path
+  code <<-EOH
+    [[ ! -d "#{fonts_path}" ]] && mkdir -m 755 -p #{fonts_path}
+    tar -C #{fonts_path} -xzf #{fonts_file} \
+    && chown root.root #{fonts_path}/* \
+    && chmod 644 #{fonts_path}/*
+    EOH
+  notifies :run, "execute[load_fonts]", :immediately
+  action :nothing
+end
+
+execute "load_fonts" do
+  command "fc-cache -fv"
+  user node[:box]['default_user']
+  action :nothing
+end